ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# RAG and Embeddings - Compatible versions
sentence-transformers==2.7.0
transformers==4.41.2
torch==2.3.1

# Database
psycopg2-binary==2.9.9

# Optional: for better performance
numpy==1.26.4